requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pandas==2.0.3
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
streamlit==1.28.0
schedule==1.2.0
transformers[torch]>=4.30.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0
torch
watchdog
--extra-index-url https://download.pytorch.org/whl/cpu